+++ tests/MapTest/CMakeLists.txt
@@ -3,9 +3,9 @@ project(Test)
 
 set(CMAKE_CXX_STANDARD 11)
 
-include_directories(../include)
+include_directories(../../include)
 
 add_executable(test
         main.cpp
-        ../src/Map.cpp
-        ../src/Position.cpp)
+        ../../src/Map.cpp
+        ../../src/Position.cpp)
